Moved oncat to databases/oncat folder
