+++ addie/processing/mantid/master_table/import_from_database/import_table_from_oncat_handler.py
@@ -7,7 +7,7 @@ import copy
 
 from addie.utilities.general import remove_white_spaces
 from addie.processing.mantid.master_table.import_from_database.oncat_template_retriever import OncatTemplateRetriever
-from addie.utilities.oncat import pyoncatGetIptsList, pyoncatGetNexus, \
+from addie.databases.oncat.oncat import pyoncatGetIptsList, pyoncatGetNexus, \
     pyoncatGetRunsFromIpts, pyoncatGetTemplate
 #from addie.processing.mantid.master_table.import_from_database import utilities as ImportFromDatabaseUtilities
 from addie.utilities.general import json_extractor
